refactor: extract chat timeline
